+++ lib/gr_blocks/nonstop_wavfile_sink.h
@@ -23,6 +23,7 @@
 #ifndef INCLUDED_GR_NONSTOP_WAVFILE_SINK_H
 #define INCLUDED_GR_NONSTOP_WAVFILE_SINK_H
 
+#include "../../trunk-recorder/global_structs.h"
 #include "../../trunk-recorder/formatter.h"
 
 //#include "../trunk-recorder/call_conventional.h"
