+++ lib/gr_blocks/nonstop_wavfile_sink_impl.h
@@ -24,6 +24,7 @@
 #define INCLUDED_GR_nonstop_wavfile_SINK_IMPL_H
 
 #include "nonstop_wavfile_sink.h"
+
 #include <sys/time.h>
 //#include "wavfile.h"
 #include <boost/log/trivial.hpp>
